CodeByOthers/HolyBot-ravi03071991/requirements.txt
pinecone-client
openai
gradio